fix: refactoring code for better maintainability
