+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './tree-key-cache';
 export * from './types';
-export { buildKey } from './utils/graphs';
-export { deserializeWholeTree, EmptyTree, TreeError } from './utils';
+export * from './utils';
