+++ src/internal/multi-tree-ref.ts
@@ -6,18 +6,24 @@ import {
 	MultiTreeValue,
 	Tree,
 	TreeKeys,
-} from './types';
-import { createTraversalItem } from './utils/graphs';
-import { getMultiValueFromTrees, isUndefined } from './utils';
+} from '../types';
+import { createTraversalItem } from '../utils/graphs';
+import { getMultiValueFromTrees } from './get-multi-value-from-trees';
+import { isUndefined } from './is-undefined';
 
+/**
+ * Tree reference representing a sync tree with history control, ie,
+ * we can have multiple values for each node
+ */
 export class MultiTreeRef<R> implements MultiTree<R> {
 	#emptyChildren = false;
 	[TreeKeys.value]?: MultiTreeValue<R>;
 	constructor(
 		private nodeRef: ChainedObject | undefined,
 		private trees: Tree<R>[],
+		private now: number,
 	) {
-		this[TreeKeys.value] = getMultiValueFromTrees(trees);
+		this[TreeKeys.value] = getMultiValueFromTrees(trees, now);
 	}
 
 	get [TreeKeys.children](): MultiTreeChildren<R> | undefined {
@@ -40,8 +46,9 @@ export class MultiTreeRef<R> implements MultiTree<R> {
 				'key',
 				({ key, trees }) =>
 					new MultiTreeRef<R>(
-						createTraversalItem<R>(key, 0, this.nodeRef, this),
+						createTraversalItem<R>(key, 0, this.nodeRef, this, this.now),
 						trees,
+						this.now,
 					),
 			);
 	}
