+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './key-tree-cache-storage';
-export * from './cache-value';
 export * from './key-tree-cache-options';
 export * from './tree-type';
