+++ src/utils/build-key.ts
@@ -1,5 +1,10 @@
 import { ChainedObject } from 'src/types';
 
+/**
+ * Builds a chained key using a ChainedObject or an array of keys
+ * @param nodeRef The chainedObject or the array of keys
+ * @returns the chained key
+ */
 export function buildKey<T extends ChainedObject>(nodeRef: T | string[]) {
 	if (Array.isArray(nodeRef)) {
 		return nodeRef.join(':');
