--- src/utils/constant.ts
@@ -1,3 +1,0 @@
-export function constant<T>(value: T) {
-	return () => value;
-}
