--- src/utils/get-multi-value-from-async-iterable.ts
@@ -1,10 +1,0 @@
-import { fluentAsync } from '@codibre/fluent-iterable';
-import { multiTreeValue } from '../types';
-
-export async function getMultiValueFromAsyncIterable<R>(
-	value: AsyncIterable<R | undefined>,
-) {
-	return {
-		[multiTreeValue]: await fluentAsync(value).filter().toArray(),
-	};
-}
