--- src/utils/get-read-storage-function.ts
@@ -1,5 +1,0 @@
-import { KeyTreeCacheStorage } from 'src/types';
-
-export function getReadStorageFunction<R>(storage: KeyTreeCacheStorage<R>) {
-	return storage.getHistory?.bind(storage) ?? storage.get.bind(storage);
-}
