+++ src/utils/graphs/create-traversal-item.ts
@@ -1,11 +1,10 @@
+import { getTreeCurrentSerializedValue } from 'src/internal';
 import {
 	AsyncTree,
 	ChainedObject,
 	MultiTree,
-	MultiTreeValue,
 	AnyTree,
 	SyncTree,
-	TreeValue,
 	Tree,
 } from '../../types';
 import {
@@ -23,48 +22,48 @@ export function createTraversalItem<T>(
 	level: number,
 	parentRef: ChainedObject | undefined,
 	treeRef: Tree<T>,
-	value?: T,
+	now: number,
 ): TraversalItem<T>;
 export function createTraversalItem<T>(
 	key: string,
 	level: number,
 	parentRef: ChainedObject | undefined,
 	treeRef: AsyncTree<T>,
-	value?: T,
+	now: number,
 ): AsyncTraversalItem<T>;
 export function createTraversalItem<T>(
 	key: string,
 	level: number,
 	parentRef: ChainedObject | undefined,
 	treeRef: MultiTree<T>,
-	value?: MultiTreeValue<T>,
+	now: number,
 ): MultiTraversalItem<T>;
 export function createTraversalItem<T>(
 	key: string,
 	level: number,
 	parentRef: ChainedObject | undefined,
 	treeRef: SyncTree<T>,
-	value?: TreeValue<T>,
+	now: number,
 ): SyncTraversalItem<T>;
 export function createTraversalItem<T>(
 	key: string,
 	level: number,
 	parentRef: ChainedObject | undefined,
 	treeRef: AnyTree<T>,
-	value?: TreeValue<T>,
+	now: number,
 ): AnyTraversalItem<T>;
 export function createTraversalItem<T>(
 	key: string,
 	level: number,
 	parentRef: ChainedObject | undefined,
 	treeRef: AnyTree<T>,
-	value?: T,
+	now: number,
 ): AnyTraversalItem<T> {
 	return {
 		key,
 		parentRef,
 		level,
-		[valueSymbol]: value,
+		[valueSymbol]: getTreeCurrentSerializedValue(treeRef, now),
 		[treeRefSymbol]: treeRef,
 	};
 }
