--- src/utils/is-undefined.ts
@@ -1,3 +1,0 @@
-export function isUndefined<T>(value: T | undefined): value is undefined {
-	return value === undefined || value === null;
-}
